Added country data formatted to publicData. Refactored utils_data into Data module. Cleaned up data methods. Added country layout component to meetups page.
