+++ public/page-data/dev-404-page/page-data.json
@@ -1,1 +1,1 @@
-{"componentChunkName":"component---cache-dev-404-page-js","path":"/dev-404-page/","result":{"data":{"allSitePage":{"nodes":[{"id":"SitePage /","path":"/"},{"id":"SitePage /meetups/","path":"/meetups/"}]}},"pageContext":{}}}
+{"componentChunkName":"component---cache-dev-404-page-js","path":"/dev-404-page/","result":{"data":{"allSitePage":{"nodes":[{"path":"/"},{"path":"/meetups/"}]}},"pageContext":{}}}
