+++ src/modules/Layouts/PageLayout.js
@@ -8,8 +8,6 @@ import Typography from "@src/styles/typography"
 import Header from "@modules/ui/Header"
 import Footer from "@modules/ui/Footer"
 
-import ThemeSwitcher from "@modules/ui/ThemeSwitcher"
-
 /**
  * PageLayout - Rejon Taylor-Foster (Feb 2, 2020)
  *
